# docs/swagger.yaml
basePath: /api
definitions:
...
  cmd.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  cmd.Filter:
    properties:
      active:
        type: boolean
      createdAt:
        type: string
      effect:
        additionalProperties: {}
        type: object
      id:
        type: string
      name:
        type: string
    type: object
  cmd.FilterCreateRequest:
    properties:
      active:
        type: boolean
      effect:
        additionalProperties: {}
        type: object
      name:
        type: string
    type: object
...
  cmd.Frame:
    properties:
      active:
        type: boolean
      createdAt:
        type: string
      fileURL:
        type: string
      id:
        type: string
      name:
        type: string
      theme:
        type: string
    type: object
  cmd.FrameCreateRequest:
    properties:
      active:
        type: boolean
      file_url:
        type: string
      name:
        type: string
      theme:
        type: string
    type: object
...
  cmd.Payment:
    properties:
      amount:
        format: float64
        type: number
      createdAt:
        type: string
      currency:
        type: string
      id:
        type: string
      method:
        $ref: '#/definitions/go-ddd-clean_internal_domain_payment.Method'
      sessionID:
        type: string
      status:
        $ref: '#/definitions/go-ddd-clean_internal_domain_payment.Status'
      transactionRef:
        type: string
    type: object
  cmd.PaymentCreateRequest:
    properties:
      amount:
        type: number
      currency:
        type: string
      method:
        type: string
      session_id:
        type: string
      status:
        type: string
      transaction_ref:
        type: string
    type: object
  cmd.PaymentUpdateRequest:
    properties:
      amount:
        type: number
      currency:
        type: string
      method:
        type: string
      status:
        type: string
      transaction_ref:
        type: string
    type: object
  cmd.Photo:
    properties:
      composition:
        additionalProperties: {}
        type: object
      createdAt:
        type: string
      filterID:
        type: string
      frameID:
        type: string
      id:
        type: string
      renderedURL:
        type: string
      sessionID:
        type: string
      storageURL:
        type: string
    type: object
  cmd.PhotoCreateRequest:
    properties:
      composition:
        additionalProperties: {}
        type: object
      filter_id:
        type: string
      frame_id:
        type: string
      rendered_url:
        type: string
      session_id:
        type: string
      storage_url:
        type: string
    type: object
...
  cmd.QRCode:
    properties:
      createdAt:
        type: string
      expireAt:
        type: string
      hash:
        type: string
      id:
        type: string
      photoID:
        type: string
    type: object
  cmd.QRCodeCreateRequest:
    properties:
      expire_at:
        type: integer
      hash:
        type: string
      photo_id:
        type: string
    type: object
...
  go-ddd-clean_internal_domain_payment.Method:
    enum:
    - cash
    - qr
    - stripe
    - points
    type: string
    x-enum-varnames:
    - MethodCash
    - MethodQR
    - MethodStripe
    - MethodPoints
  go-ddd-clean_internal_domain_payment.Status:
    enum:
    - pending
    - success
    - failed
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusSuccess
    - StatusFailed
...
info:
  contact: {}
  description: API documentation for the go-ddd-clean service.
  title: Photobooth Platforms API
  version: "1.0"
paths:
...
  /api/media/filters:
    get:
      parameters:
      - description: ดึงเฉพาะที่เปิดใช้งาน
        in: query
        name: active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/cmd.Filter'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      summary: ดึงรายการฟิลเตอร์
      tags:
      - Media Filters
    post:
      consumes:
      - application/json
      parameters:
      - description: ข้อมูลฟิลเตอร์
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/cmd.FilterCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/cmd.Filter'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      summary: สร้างฟิลเตอร์
      tags:
      - Media Filters
...
  /api/media/frames:
    get:
      parameters:
      - description: ดึงเฉพาะที่เปิดใช้งาน
        in: query
        name: active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/cmd.Frame'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      summary: ดึงรายการกรอบรูป
      tags:
      - Media Frames
    post:
      consumes:
      - application/json
      parameters:
      - description: ข้อมูลกรอบรูป
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/cmd.FrameCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/cmd.Frame'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      summary: สร้างกรอบรูป
      tags:
      - Media Frames
...
  /api/media/photos:
    get:
      parameters:
      - description: รหัสเซสชัน
        in: query
        name: session_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/cmd.Photo'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      security:
      - BoothTokenAuth: []
      summary: ดึงรายการรูป
      tags:
      - Media Photos
    post:
      consumes:
      - application/json
      parameters:
      - description: ข้อมูลรูป
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/cmd.PhotoCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/cmd.Photo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      security:
      - BoothTokenAuth: []
      summary: สร้างรูปใหม่
      tags:
      - Media Photos
...
  /api/media/qrcodes:
    post:
      consumes:
      - application/json
      parameters:
      - description: ข้อมูล QR Code
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/cmd.QRCodeCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/cmd.QRCode'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      summary: สร้าง QR Code
      tags:
      - Media QRCodes
...
  /api/payments:
    post:
      consumes:
      - application/json
      parameters:
      - description: ข้อมูลการชำระเงิน
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/cmd.PaymentCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/cmd.Payment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      summary: สร้างข้อมูลการชำระเงิน
      tags:
      - Payments
  /api/payments/{id}:
    get:
      parameters:
      - description: รหัสการชำระเงิน
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd.Payment'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      summary: ดูข้อมูลการชำระเงิน
      tags:
      - Payments
    put:
      consumes:
      - application/json
      parameters:
      - description: รหัสการชำระเงิน
        in: path
        name: id
        required: true
        type: string
      - description: ข้อมูลที่ต้องแก้ไข
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/cmd.PaymentUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd.Payment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      summary: ปรับปรุงข้อมูลการชำระเงิน
      tags:
      - Payments
  /api/payments/session/{sessionID}:
    get:
      parameters:
      - description: รหัสเซสชัน
        in: path
        name: sessionID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd.Payment'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      summary: ดูข้อมูลการชำระเงินของเซสชัน
      tags:
      - Payments
...
  /api/sessions/{id}/payment:
    get:
      parameters:
      - description: รหัสเซสชัน
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd.Payment'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/cmd.ErrorResponse'
      security:
      - BoothTokenAuth: []
      summary: ดูข้อมูลการชำระเงินของเซสชัน
      tags:
      - Sessions
...
schemes:
- http
securityDefinitions:
  BoothTokenAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"